+++ packages/dnd/src/isTouchDevice.ts
@@ -1,0 +1,1 @@
+export const isTouchDevice = (): boolean => !!window?.ontouchstart || navigator?.maxTouchPoints > 0;